Modules practice.
this is a TOP project where they invite you to practice
having as little global code as possible. i probalby overdid it
i mean just look at the main file which is empty but it did find
it much cleaner to work with modules. let's hope that i can revisit
this project and have it make sense.